backend/requirements.txt
# FastAPI e servidor
fastapi==0.104.1
uvicorn[standard]==0.24.0
python-multipart==0.0.6

# Banco de dados
sqlalchemy==2.0.23
alembic==1.12.1
psycopg2-binary==2.9.9

# Validação e serialização
pydantic==2.5.0
pydantic-settings==2.1.0

# Autenticação e segurança
python-jose[cryptography]==3.3.0
passlib[bcrypt]==1.7.4
python-multipart==0.0.6

# QR Code
qrcode[pil]==7.4.2
Pillow==10.1.0

# Cache e locks
redis==5.0.1

# Relatórios
pandas==2.1.3
openpyxl==3.1.2
reportlab==4.0.7

# Utilitários
python-dotenv==1.0.0
email-validator==2.1.0

# Testes
pytest==7.4.3
pytest-asyncio==0.21.1
httpx==0.25.2



